Create Personal Finance Calculator

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expense = int(input("Enter your monthly expense: "))
+monthly_income = float(input("Enter your monthly income: "))
+monthly_expense = float(input("Enter your monthly expense: "))
 
 monthly_savings = monthly_income - monthly_expense
 annual_interest = 5/100
